Update host_interfaces.j2
The ifup -a command in ~/automation/roles/anycast_host/tasks/main.yml hang. The problem seems to be that it is trying to bring up the slave interfaces eth1 and eth2 before the master bond0 interface is up. Moving the bond0 stanza ahead of those for eth1 and eth2 and fixes it.